Clean up Options
